fix: align BSDIFF buffer length types

+++ app/android/app/src/main/kotlin/com/daidai/daidai_app/AndroidBsPatch.kt
@@ -53,7 +53,7 @@ object AndroidBsPatch {
                                         val prefix = (readableStart - oldPosition).coerceAtMost(count.toLong()).toInt()
                                         if (prefix < count && readableStart < old.length()) {
                                             old.seek(readableStart)
-                                            val available = minOf(count - prefix, old.length() - readableStart).toInt()
+                                            val available = minOf((count - prefix).toLong(), old.length() - readableStart).toInt()
                                             old.readFully(oldBuffer, prefix, available)
                                         }
                                         repeat(count) { index ->
